Improve how pages/outlines are handled when narrow
1. If logos-outlines-are-pages is non-nil, it now includes the match of
the page-delimiter regexp in the narrowed region and leaves point right
after the page-delimiter ---so in Org mode, after the stars.  (I thought
it was better to leave point there than at the very beginning of the
narrowed buffer to match the behavior of logos-forward-page-dwim when
the buffer is not narrowed.)

2. To avoid skipping pages in the narrowed case, the function checks if
you are right at the start of a page-delimiter and if so move you past
the delimiter in the opposite direction you are moving: so if you are
moving back, it puts you after the delimiter, and if you are moving
forward it puts you before the delimiter.  (The bug was that if the
point was at point-max while narrowed and moving forward, it would skip
past a page and the same in the opposite direction with point-min.)

3. Changed logos-narrow-dwim to call logos--narrow-to-page instead of
narrow-to-page, so that it too includes the page-delimiter match in the
page.