Proposal about elisp code vs customization menu
The variables that can be set in the customization menu, maybe shouldn't appear as elisp code in the section "4 Sample Configuration". Because it is possible that the user copies that configuration from the elisp code into the init file, and when she tries to customize logos using the customization menu (maybe months later), the changes in the customization menu are "crushed" by the previously pasted elisp code.

I haven't made all the changes, just an example of my proposal.