Guard against void outline-regexp
The previous value would cause a problem if the outline-regexp had not
been set yet.  In a typical Emacs session, even on 'emacs -Q', the
outline-regexp is set in the scratch buffer, so we always have a valid
value.  However, when the user changes the initial-buffer-choice or
opens Emacs directly for a given file, the outline-regexp can be void,
causing an error with logos-outline-regexp-alist.

Thanks to Xiaoduan for informing me about this error and for testing my
code on how to fix it.  This was done via a private channel and is
shared with permission.