REVISE fallback of logos-outline-regexp-alist
The previous value did not account for the buffer-local values of
'outline-regexp' or 'page-delimiter', nor did it adapt to any changes
in those values.  Instead, it would capture whatever the value was at
the time of declaration.

We remove the 't' condition altogether and instead handle the fallback
value internally.  Logos will either return the current buffer-local
value of 'outline-regexp' or 'page-delimiter', in this order.

Put simply, this has the same intent as before, but works properly.